Update namespace, update env reference. (#193)
