--- responsibleai/rai/components/causal/spec.yaml
@@ -1,83 +1,0 @@
-$schema: http://azureml/sdk-2-0/CommandComponent.json
-name: rai_insights_causal
-display_name: Add Causal to RAI Insights Dashboard (Private preview)
-description: Add Causal to RAI Insights Dashboard [Learn More](https://github.com/Azure/RAI-vNext-Preview)
-version: 0.0.1-preview
-type: command
-
-inputs:
-  rai_insights_dashboard:
-    type: path
-  treatment_features:
-    type: string # JSON encoded list
-  heterogeneity_features:
-    type: string # Optional JSON encoded list
-    default: 'null'
-  nuisance_model:
-    type: string # Enum
-    default: linear
-    enum: ['linear', 'automl']
-  heterogeneity_model:
-    type: string # Enum
-    default: linear
-    enum: ['linear', 'forest']
-  alpha:
-    type: number
-    default: 0.05
-  upper_bound_on_cat_expansion:
-    type: integer
-    default: 50
-  treatment_cost:
-    type: string
-    default: '0' # Float or JSON encoded list of floats
-  min_tree_leaf_samples:
-    type: integer
-    default: 2
-  max_tree_depth:
-    type: integer
-    default: 2
-  skip_cat_limit_checks:
-    type: boolean
-    default: False
-  categories:
-    type: string
-    default: auto
-  n_jobs:
-    type: integer
-    default: 1
-  verbose:
-    type: integer
-    default: 1
-  random_state:
-    type: string # int or none
-    default: None
-
-outputs:
-  causal:
-    type: path
-
-code: ../src/
-
-environment: azureml:AzureML-responsibleai-0.18-ubuntu20.04-py38-cpu@latest
-
-# For debugging
-environment_variables:
-   AZUREML_COMPUTE_USE_COMMON_RUNTIME: true
-
-command: >-
-  python create_causal.py
-  --rai_insights_dashboard ${{inputs.rai_insights_dashboard}}
-  --treatment_features '${{inputs.treatment_features}}'
-  --heterogeneity_features '${{inputs.heterogeneity_features}}'
-  --nuisance_model ${{inputs.nuisance_model}}
-  --heterogeneity_model ${{inputs.heterogeneity_model}}
-  --alpha ${{inputs.alpha}}
-  --upper_bound_on_cat_expansion ${{inputs.upper_bound_on_cat_expansion}}
-  --treatment_cost '${{inputs.treatment_cost}}'
-  --min_tree_leaf_samples ${{inputs.min_tree_leaf_samples}}
-  --max_tree_depth ${{inputs.max_tree_depth}}
-  --categories ${{inputs.categories}}
-  --n_jobs ${{inputs.n_jobs}}
-  --verbose ${{inputs.verbose}}
-  --random_state '${{inputs.random_state}}'
-  --causal_path ${{outputs.causal}}
