--- responsibleai/rai/components/counterfactual/spec.yaml
@@ -1,53 +1,0 @@
-$schema: http://azureml/sdk-2-0/CommandComponent.json
-name: rai_insights_counterfactual
-display_name: Add Counterfactuals to RAI Insights Dashboard (Private preview)
-description: Add Counterfactuals to RAI Insights Dashboard [Learn More](https://github.com/Azure/RAI-vNext-Preview)
-version: 0.0.1-preview
-type: command
-
-inputs:
-  rai_insights_dashboard:
-    type: path
-  total_CFs:
-    type: integer
-    default: 10
-  method:
-    type: string # Enum
-    default: random
-    enum: ['random', 'genetic', 'kdtree']
-  desired_class:
-    type: string # Or int
-    default: 0
-  desired_range:
-    type: string # JSON encoded list
-    default: '[]'
-  permitted_range:
-    type: string # JSON encoded dict
-    default: '{}'
-  features_to_vary:
-    type: string # Or JSON encoded list
-    default: all
-  feature_importance:
-    type: boolean
-    default: True
-
-
-outputs:
-  counterfactual:
-    type: path
-
-code: ../src/
-
-environment: azureml:AzureML-responsibleai-0.18-ubuntu20.04-py38-cpu@latest
-
-command: >-
-  python create_counterfactual.py
-  --rai_insights_dashboard ${{inputs.rai_insights_dashboard}}
-  --total_CFs ${{inputs.total_CFs}}
-  --method ${{inputs.method}}
-  --desired_class ${{inputs.desired_class}}
-  --desired_range '${{inputs.desired_range}}'
-  --permitted_range '${{inputs.permitted_range}}'
-  --features_to_vary '${{inputs.features_to_vary}}'
-  --feature_importance '${{inputs.feature_importance}}'
-  --counterfactual_path ${{outputs.counterfactual}}
