--- responsibleai/rai/components/insight_create/spec.yaml
@@ -1,46 +1,0 @@
-$schema: http://azureml/sdk-2-0/CommandComponent.json
-name: rai_insights_constructor
-display_name: RAI Insights Dashboard Constructor (Private preview)
-description: RAI Insights Dashboard Constructor [Learn More](https://github.com/Azure/RAI-vNext-Preview)
-version: 0.0.1-preview
-type: command
-inputs:
-  title:
-    type: string
-  task_type:
-    type: string # [classification, regression]
-    enum: ['classification', 'regression']
-  model_info_path:
-    type: path # To model_info.json
-  train_dataset:
-    type: path # Must be Parquet
-  test_dataset:
-    type: path # Must be Parquet
-  target_column_name:
-    type: string
-  maximum_rows_for_test_dataset:
-    type: integer
-    default: 5000
-  categorical_column_names:
-    type: string # Optional[List[str]]
-    default: '[]'
-  classes:
-    type: string # Optional[List[str]]
-    default: '[]'
-outputs:
-  rai_insights_dashboard:
-    type: path
-code: ../src/
-environment: azureml:AzureML-responsibleai-0.18-ubuntu20.04-py38-cpu@latest
-command: >-
-  python create_rai_insights.py
-  --title '${{inputs.title}}'
-  --task_type ${{inputs.task_type}}
-  --model_info_path ${{inputs.model_info_path}}
-  --train_dataset ${{inputs.train_dataset}}
-  --test_dataset ${{inputs.test_dataset}}
-  --target_column_name ${{inputs.target_column_name}}
-  --maximum_rows_for_test_dataset ${{inputs.maximum_rows_for_test_dataset}}
-  --categorical_column_names '${{inputs.categorical_column_names}}'
-  --classes '${{inputs.classes}}'
-  --output_path ${{outputs.rai_insights_dashboard}}
