--- responsibleai/rai/components/score_card/spec.yaml
@@ -1,29 +1,0 @@
-$schema: http://azureml/sdk-2-0/CommandComponent.json
-name: rai_score_card
-display_name: Generate rai insight score card (Private preview)
-description: Generate rai insight score card [Learn More](https://github.com/Azure/RAI-vNext-Preview)
-version: 0.0.1-preview
-type: command
-
-inputs:
-  dashboard:
-    type: path
-  pdf_generation_config:
-    type: path
-  predefined_cohorts_json:
-    type: path
-    optional: true
-outputs:
-  scorecard:
-    type: path
-
-environment: azureml:AzureML-responsibleai-0.18-ubuntu20.04-py38-cpu@latest
-
-code: ../src/
-
-command: >-
-  python create_score_card.py
-  --rai_insights_dashboard ${{inputs.dashboard}}
-  --pdf_generation_config ${{inputs.pdf_generation_config}}
-  [--predefined_cohorts_json ${{inputs.predefined_cohorts_json}}]
-  --pdf_output_path ${{outputs.scorecard}}
