--- responsibleai/rai/components/src/create_explanation.py
@@ -1,78 +1,0 @@
-# Copyright (c) Microsoft Corporation.
-# Licensed under the MIT License.
-
-import argparse
-import logging
-
-from responsibleai import RAIInsights
-
-from azureml.core import Run
-
-from constants import RAIToolType
-from rai_component_utilities import (
-    create_rai_insights_from_port_path,
-    save_to_output_port,
-    copy_dashboard_info_file,
-)
-
-_logger = logging.getLogger(__file__)
-logging.basicConfig(level=logging.INFO)
-
-
-def parse_args():
-    # setup arg parser
-    parser = argparse.ArgumentParser()
-
-    parser.add_argument("--rai_insights_dashboard", type=str, required=True)
-    parser.add_argument("--comment", type=str, required=True)
-    parser.add_argument("--explanation_path", type=str, required=True)
-
-    # parse args
-    args = parser.parse_args()
-
-    # return args
-    return args
-
-
-def main(args):
-
-    my_run = Run.get_context()
-
-    # Create the RAI Insights object
-    rai_i: RAIInsights = create_rai_insights_from_port_path(
-        my_run, args.rai_insights_dashboard
-    )
-
-    # Add the explanation
-    rai_i.explainer.add()
-    _logger.info("Added explanation")
-
-    # Compute
-    rai_i.compute()
-    _logger.info("Computation complete")
-
-    # Save
-    save_to_output_port(rai_i, args.explanation_path, RAIToolType.EXPLANATION)
-    _logger.info("Saved to output port")
-
-    # Copy the dashboard info file
-    copy_dashboard_info_file(args.rai_insights_dashboard, args.explanation_path)
-
-    _logger.info("Completing")
-
-
-# run script
-if __name__ == "__main__":
-    # add space in logs
-    print("*" * 60)
-    print("\n\n")
-
-    # parse args
-    args = parse_args()
-
-    # run main function
-    main(args)
-
-    # add space in logs
-    print("*" * 60)
-    print("\n\n")
