+++ responsibleai/tabular/components/error_analysis/spec.yaml
@@ -1,0 +1,39 @@
+$schema: http://azureml/sdk-2-0/CommandComponent.json
+name: microsoft_azureml_rai_tabular_erroranalysis
+display_name: Add Error Analysis to RAI Insights Dashboard (Private preview)
+description: Add Error Analysis to RAI Insights Dashboard [Learn More](https://github.com/Azure/RAI-vNext-Preview)
+version: 0.0.1-preview
+type: command
+
+inputs:
+  rai_insights_dashboard:
+    type: path
+  max_depth:
+    type: integer
+    default: 3
+  num_leaves:
+    type: integer
+    default: 31
+  min_child_samples:
+    type: integer
+    default: 20
+  filter_features:
+    type: string
+    default: 'null'
+
+outputs:
+  error_analysis:
+    type: path
+
+code: ../src/
+
+environment: azureml://registries/azureml-dev/environments/AzureML-responsibleai-0.18-ubuntu20.04-py38-cpu/versions/10
+
+command: >-
+  python create_error_analysis.py
+  --rai_insights_dashboard ${{inputs.rai_insights_dashboard}}
+  --max_depth ${{inputs.max_depth}}
+  --num_leaves ${{inputs.num_leaves}}
+  --min_child_samples ${{inputs.min_child_samples}}
+  --filter_features '${{inputs.filter_features}}'
+  --error_analysis_path ${{outputs.error_analysis}}
