+++ responsibleai/tabular/components/explanation/spec.yaml
@@ -1,0 +1,26 @@
+$schema: http://azureml/sdk-2-0/CommandComponent.json
+name: microsoft_azureml_rai_tabular_explanation
+display_name: Add Explanation to RAI Insights Dashboard (Private preview)
+description: Add Explanation to RAI Insights Dashboard [Learn More](https://github.com/Azure/RAI-vNext-Preview)
+version: 0.0.1-preview
+type: command
+
+inputs:
+  comment:
+    type: string
+  rai_insights_dashboard:
+    type: path
+
+outputs:
+  explanation:
+    type: path
+
+code: ../src/
+
+environment: azureml://registries/azureml-dev/environments/AzureML-responsibleai-0.18-ubuntu20.04-py38-cpu/versions/10
+
+command: >-
+  python create_explanation.py
+  --comment '${{inputs.comment}}'
+  --rai_insights_dashboard ${{inputs.rai_insights_dashboard}}
+  --explanation_path ${{outputs.explanation}}
