+++ responsibleai/tabular/components/insight_gather/spec.yaml
@@ -1,0 +1,42 @@
+$schema: http://azureml/sdk-2-0/CommandComponent.json
+name: microsoft_azureml_rai_tabular_insight_gather
+display_name: Gather RAI Insights Dashboard (Private preview)
+description: Gather RAI Insights Dashboard [Learn More](https://github.com/Azure/RAI-vNext-Preview)
+version: 0.0.1-preview
+type: command
+
+inputs:
+  constructor:
+    type: path
+  insight_1:
+    type: path
+    optional: true
+  insight_2:
+    type: path
+    optional: true
+  insight_3:
+    type: path
+    optional: true
+  insight_4:
+    type: path
+    optional: true
+
+outputs:
+  dashboard:
+    type: path
+  ux_json:
+    type: path
+
+environment: azureml://registries/azureml-dev/environments/AzureML-responsibleai-0.18-ubuntu20.04-py38-cpu/versions/10
+
+code: ../src/
+
+command: >-
+  python gather_rai_insights.py
+  --constructor ${{inputs.constructor}}
+  $[[--insight_1 '${{inputs.insight_1}}']]
+  $[[--insight_2 '${{inputs.insight_2}}']]
+  $[[--insight_3 '${{inputs.insight_3}}']]
+  $[[--insight_4 '${{inputs.insight_4}}']]
+  --dashboard ${{outputs.dashboard}}
+  --ux_json ${{outputs.ux_json}}
