+++ responsibleai/tabular/components/src/_score_card/box_plot_explain.png
