+++ responsibleai/tabular/jobs/boston_analyze_scorecard_config/pdf_gen.json
@@ -1,0 +1,28 @@
+{
+  "Model": {
+    "ModelName": "Boston Analyze",
+    "ModelType": "Regression",
+    "ModelSummary": "This is a regression model for boston analyze"
+  },
+  "Metrics": {
+    "mean_absolute_error": {
+      "threshold": "<=20"
+    },
+    "mean_squared_error": {}
+  },
+  "FeatureImportance": {
+    "top_n": 6
+  },
+  "DataExplorer": {
+    "features": [
+      "TAX",
+      "AGE",
+      "CRIM",
+      "DIS"
+    ]
+  },
+  "Cohorts": [
+    "High Tax",
+    "Low Tax"
+  ]
+}
