+++ responsibleai/tabular/jobs/data/boston_test/MLTable
@@ -1,0 +1,6 @@
+$schema: http://azureml/sdk-2-0/MLTable.json
+type: mltable
+paths:
+  - file: ./boston_test.parquet
+transformations:
+  - read_parquet
