+++ responsibleai/tabular/jobs/pipeline_boston_analyze.yml
@@ -1,0 +1,114 @@
+$schema: https://azuremlschemas.azureedge.net/latest/pipelineJob.schema.json
+experiment_name: boston_analyze
+type: pipeline
+
+inputs:
+  target_column_name: 'y'
+  my_training_data:
+    type: mltable
+    path: ./data/boston_train
+  my_test_data:
+    type: mltable
+    path: ./data/boston_test
+
+outputs: 
+  boston_model:
+    type: mlflow_model
+    mode: rw_mount
+
+compute: azureml:cpucluster
+
+jobs:
+  train_model:
+    type: command
+    component: file:./train_boston.yml
+    inputs:
+      training_data: ${{parent.inputs.my_training_data}}
+      target_column_name: ${{parent.inputs.target_column_name}}
+      categorical_features: '[]'
+      continuous_features: '["CRIM", "ZN", "INDUS", "CHAS", "NOX", "RM", "AGE","DIS", "RAD", "TAX", "PTRATIO", "B", "LSTAT"]'
+      model_name: 'boston_model'
+    outputs:
+      model_output: ${{parent.outputs.boston_model}}
+
+  create_rai_job:
+    type: command
+    component: file:../components/insight_create/spec.yaml
+    limits:
+      timeout: 120
+    inputs:
+      title: Boston Housing Analysis
+      task_type: regression
+      model_info_path: ${{parent.jobs.train_model.outputs.model_output_json}}
+      train_dataset: ${{parent.inputs.my_training_data}}
+      test_dataset: ${{parent.inputs.my_test_data}}
+      target_column_name: ${{parent.inputs.target_column_name}}
+      categorical_column_names: '[]'
+
+  explain_01:
+    type: command
+    component: file:../components/explanation/spec.yaml
+    limits:
+      timeout: 120
+    inputs:
+      comment: Some random string
+      rai_insights_dashboard: ${{parent.jobs.create_rai_job.outputs.rai_insights_dashboard}}
+
+  causal_01:
+    type: command
+    component: file:../components/causal/spec.yaml
+    limits:
+      timeout: 120
+    inputs:
+      rai_insights_dashboard: ${{parent.jobs.create_rai_job.outputs.rai_insights_dashboard}}
+      treatment_features: '["ZN", "NOX"]'
+      heterogeneity_features: '[]'
+      nuisance_model: linear
+      skip_cat_limit_checks: True
+
+  counterfactual_01:
+    type: command
+    component: file:../components/counterfactual/spec.yaml
+    limits:
+      timeout: 600
+    inputs:
+      rai_insights_dashboard: ${{parent.jobs.create_rai_job.outputs.rai_insights_dashboard}}
+      total_CFs: 10
+      desired_range: '[10, 300]'
+      feature_importance: True
+
+  error_analysis_01:
+    limits:
+      timeout: 120
+    type: command
+    component: file:../components/error_analysis/spec.yaml
+    inputs:
+      rai_insights_dashboard: ${{parent.jobs.create_rai_job.outputs.rai_insights_dashboard}}
+      max_depth: 3
+      filter_features: '[]'
+
+  gather_01:
+    type: command
+    component: file:../components/insight_gather/spec.yaml
+    limits:
+      timeout: 120
+    inputs:
+      constructor: ${{parent.jobs.create_rai_job.outputs.rai_insights_dashboard}}
+      insight_1: ${{parent.jobs.causal_01.outputs.causal}}
+      insight_2: ${{parent.jobs.counterfactual_01.outputs.counterfactual}}
+      insight_3: ${{parent.jobs.error_analysis_01.outputs.error_analysis}}
+      insight_4: ${{parent.jobs.explain_01.outputs.explanation}}
+
+  scorecard_01:
+    type: command
+    component: file:../components/score_card/spec.yaml
+    inputs:
+      dashboard: ${{parent.jobs.gather_01.outputs.dashboard}}
+      pdf_generation_config:
+        type: uri_file
+        path: ./boston_analyze_scorecard_config/pdf_gen.json
+        mode: download
+      predefined_cohorts_json:
+        type: uri_file
+        path: ./boston_analyze_scorecard_config/cohorts.json
+        mode: download
