+++ responsibleai/validation_rules.json
@@ -1,5 +1,8 @@
 {
   "pep8": {
     "max-line-length": 200
+  },
+  "doc": {
+    "exclude": "./"
   }
 }
